Consolodate PacketDelayStats in one place; populate it for old octo.
